add phone validator
