+++ .idea/encodings.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/services/api-gateway/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/config-server/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/services/customer/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/services/customer-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/discovery-server/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/services/gateway/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/notification/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/order/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/payment/src/main/java" charset="UTF-8" />
