--- services/api-gateway/src/main/resources/application.yml
@@ -1,5 +1,0 @@
-spring:
-  application:
-    name: api-gateway
-  config:
-    import: optional:configserver:http://localhost:8888
