--- services/config-server/src/main/resources/configurations/application.yml
@@ -1,17 +1,0 @@
-eureka:
-  instance:
-    hostname: localhost
-  client:
-    service-url:
-      defaultZone: http://localhost:8761/eureka
-name:
-  value: david
-spring:
-  cloud:
-    config:
-      override-system-properties: false
-
-management:
-  tracing:
-    sampling:
-      probability: 1.0
