+++ services/customer-service/src/main/java/com/davidnguyenshop/app/annotation/CountryEnums.java
@@ -1,0 +1,6 @@
+package com.davidnguyenshop.app.annotation;
+
+public enum CountryEnums {
+    VN,
+    US
+}
