+++ services/customer-service/src/main/java/com/davidnguyenshop/app/annotation/PhoneValidator.java
@@ -1,0 +1,23 @@
+package com.davidnguyenshop.app.annotation;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Constraint(validatedBy = PhoneValidatorImpl.class)
+@Target({ElementType.FIELD, ElementType.PARAMETER})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface PhoneValidator {
+
+    String country() default "VN";
+
+    String message() default "Invalid phone number";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+}
