+++ services/customer-service/src/main/java/com/davidnguyenshop/app/utils/CustomerUtils.java
