+++ services/customer-service/src/main/resources/locale/message_vi.properties
